Update cell-by-cell.md
- Corrected the info for the "getting all the same color one on side" method

Added much more thorough information for OLC

If the pull request is accepted, I should probably remove the info that is in the "Last Layer Strategy" Page. If the new OLC info is better suited for that page instead of the Cell by Cell page, I can move the info to that page if needed.